Move getTermsLink function from SignDisclaimer to /InsuranceValues/index.
